app/ringier-a2a-sdk/pyproject.toml
[project]
name = "ringier-a2a-sdk"
version = "0.1.0"
description = "Ringier A2A SDK for building authenticated agent-to-agent communication systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyjwt>=2.10.1",
    "aiohttp>=3.11.0",
    "httpx>=0.28.1",
    "starlette>=0.46.2",
    "a2a-sdk>=0.3.9",
    "cryptography>=43.0.0",
    "authlib>=1.6.5",
]

[project.optional-dependencies]
langgraph = [
    "langchain>=0.3.18",
    "langchain-aws>=0.3.0",
    "langgraph>=0.2.72",
    "langgraph-checkpoint-aws>=0.1.4",
    "langchain-mcp-adapters>=0.0.6",
    "deepagents>=0.2.0",
    "boto3>=1.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    # Include langgraph dependencies for comprehensive testing
    "langchain>=0.3.18",
    "langchain-aws>=0.3.0",
    "langgraph>=0.2.72",
    "langgraph-checkpoint-aws>=0.1.4",
    "langchain-mcp-adapters>=0.0.6",
    "deepagents>=0.2.0",
    "boto3>=1.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
